% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preTeste.R
\name{pre.teste}
\alias{pre.teste}
\title{Funcao pre.teste() para sugestao de teste estatistico}
\usage{
pre.teste(..., alpha = 0.05, ajuda = FALSE, verbose = TRUE)
}
\arguments{
\item{...}{Dois ou mais vetores (numericos ou categoricos)}

\item{alpha}{Nivel de significancia. Padrao = 0.05}

\item{ajuda}{Logico. Se TRUE, mostra ajuda detalhada}

\item{verbose}{Logico. Se TRUE, imprime mensagens informativas}
}
\value{
Lista invisivel com resultados dos testes de normalidade, homogeneidade ou tabela de contingencia e recomendacao do teste
}
\description{
Identifica automaticamente se os dados sao numericos ou categoricos e sugere o teste estatistico mais adequado.
}
